Move some CQRS infrastructure to Backend.Core
