+++ Backend.WebApi/Cqrs/ExternalAccountHandler.cs
@@ -1,4 +1,5 @@
 using Backend.Core.Authentication;
+using Backend.Core.Cqrs.Infrastructure;
 using Backend.EmailIngestion.Services;
 using Backend.WebApi.Cqrs.Infrastructure;
