+++ Backend.WebApi/Cqrs/Infrastructure/DispatchController.cs
@@ -1,5 +1,6 @@
-using System.Text.Json;
 using System.ComponentModel;
+using System.Text.Json;
+using Backend.Core.Cqrs.Infrastructure;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
